ai/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
audioread==3.0.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
datasets==2.17.1
decorator==4.4.2
dill==0.3.8
evaluate==0.4.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.16.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
grpcio-tools==1.60.1
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
stable-ts==2.15.6
sympy==1.12
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
